+++ libmarch/include/march/python/wrapper_march.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+/*
+ * Copyright (c) 2017, Yung-Yu Chen <yyc@solvcon.net>
+ * BSD 3-Clause License, see COPYING
+ */
+
+#include <pybind11/pybind11.h>
+
+#include "march/python/WrapBase.hpp"
+#include "march/python/wrapper_core.hpp"
+#include "march/python/wrapper_mesh.hpp"
+
+// vim: set ff=unix fenc=utf8 nobomb et sw=4 ts=4:
